Refactoring

+++ CryptoScanner.Config/ViewModels/AnalyzerCommonViewModel.cs
@@ -1,7 +1,6 @@
 using CommunityToolkit.Mvvm.ComponentModel;
 
 using CryptoScanner.Core.Core;
-using CryptoScanner.Core.Enums;
 using CryptoScanner.Core.Settings;
 
 namespace CryptoScanner.Config.ViewModels;
@@ -10,17 +9,17 @@ public partial class AnalyzerCommonViewModel : ObservableObject
 {
     // Check relative change% 24 hours (groupBox9)
     [ObservableProperty]
-    private double _analysisMinChangePercentage = 0;
+    private float _analysisMinChangePercentage = 0;
 
     [ObservableProperty]
-    private double _analysisMaxChangePercentage = 25;
+    private float _analysisMaxChangePercentage = 25;
 
     [ObservableProperty]
     private bool _logAnalysisMinMaxChangePercentage = false;
 
     // Check effective change% over multiple days (GroupBoxXDaysEffective)
     [ObservableProperty]
-    private double _analysisEffectivePercentage = 25;
+    private float _analysisEffectivePercentage = 25;
 
     [ObservableProperty]
     private int _analysisEffectiveDays = 1;
@@ -52,7 +51,7 @@ public partial class AnalyzerCommonViewModel : ObservableObject
     private int _removeSignalAfterxCandles = 15;
 
     [ObservableProperty]
-    private bool _showInvalidSignals  = false;
+    private bool _showInvalidSignals = false;
 
     // Fine tuning (later)
     [ObservableProperty]
