+++ CryptoScanner.Core/Exchange/Bitvavo/Spot/Symbol.cs
@@ -23,7 +23,7 @@ public class Symbol() : SymbolBase(), ISymbol
                 var api = client.SpotApi;
                 using CryptoDatabase database = new();
                 database.Open();
-               
+
 
                 // Tickers for the 24h volume
                 GlobalData.AddTextToLogTab($"Reading symbol ticker information from {ExchangeBase.ExchangeOptions.ExchangeName}");
@@ -91,7 +91,7 @@ public class Symbol() : SymbolBase(), ISymbol
 
                                 // volume from the tickers
                                 if (volumeTicker.TryGetValue(symbol.Name, out decimal volume))
-                                    symbol.Volume = volume;
+                                    symbol.Volume = (double)volume;
                                 else
                                     symbol.Volume = 0;
