+++ CryptoScanner.Core/Exchange/HyperLiquid/Spot/Symbol.cs
@@ -111,7 +111,7 @@ public class Symbol() : SymbolBase(), ISymbol
 
                                 // volume from the tickers
                                 if (volumeTicker.TryGetValue(symbol.Name, out decimal volume))
-                                    symbol.Volume = volume;
+                                    symbol.Volume = (double)volume;
                                 else
                                     symbol.Volume = 0;
