+++ CryptoScanner.Core/Exchange/Kucoin/Spot/Symbol.cs
@@ -79,7 +79,7 @@ public class Symbol() : SymbolBase(), ISymbol
                                 SymbolInfo info = ParseSymbol(symbolData.Symbol, symbolData.BaseAsset, symbolData.QuoteAsset);
                                 if (IsSymbolAccepted(exchange, info, api, TradingMode.Spot, out CryptoSymbol? symbol))
                                 {
-                                      //Tijdelijk alles overnemen (vanwege into nieuwe velden)
+                                    //Tijdelijk alles overnemen (vanwege into nieuwe velden)
                                     //De te gebruiken precisie in prijzen
                                     //symbol.BaseAssetPrecision = binanceSymbol.LotSizeFilter.BasePrecision.ToString().Length - 2;
                                     //if (symbol.BaseAssetPrecision <= 0)
@@ -108,7 +108,7 @@ public class Symbol() : SymbolBase(), ISymbol
 
                                     // volume from the tickers
                                     if (volumeTicker.TryGetValue(symbol.Name, out decimal volume))
-                                        symbol.Volume = volume;
+                                        symbol.Volume = (double)volume;
                                     else
                                         symbol.Volume = 0;
