+++ CryptoScanner.Core/Settings/SettingsSignal.cs
@@ -22,13 +22,13 @@ public class SettingsSignal
     public bool LogMinimumTickPercentage { get; set; } = false;
 
     // de 24 change moet binnen dit interval zitten (start/end price)
-    public double AnalysisMinChangePercentage { get; set; } = -25;
-    public double AnalysisMaxChangePercentage { get; set; } = 25;
+    public float AnalysisMinChangePercentage { get; set; } = -25;
+    public float AnalysisMaxChangePercentage { get; set; } = 25;
     public bool LogAnalysisMinMaxChangePercentage { get; set; } = false;
 
     // de x dagen effectief moet binnen dit interval zitten (full effective move)
     public int AnalysisEffectiveDays { get; set; } = 5;
-    public double AnalysisEffectivePercentage { get; set; } = 35;
+    public float AnalysisEffectivePercentage { get; set; } = 35;
     public bool AnalysisMaxEffectiveLog { get; set; } = false;
 
     // STOBB signals
@@ -50,7 +50,7 @@ public class SettingsSignal
     public SettingsSignalStrategyFvg ZonesFvg = new();
 
     // Nadaraya Watson Envelope
-    public SettingsSignalStrategyNwe Nwe = new ();
+    public SettingsSignalStrategyNwe Nwe = new();
 
 
     // Logging
